change MainView to HomeFragment
