+++ MountainWalker.Droid/Resources/layout/MainView.axml
@@ -11,26 +11,6 @@
         android:clickable="true">
         <include
             layout="@layout/Toolbar" />
-        
-        <fragment
-            android:id="@+id/map"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            class="com.google.android.gms.maps.MapFragment">
-            <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Start"
-                android:layout_gravity="center_horizontal"
-                local:MvxBind="Click OpenMainDialogCommand" />
-            <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Logout"
-                android:layout_gravity="bottom|left"
-                local:MvxBind="Click LogoutCommand" />
-        </fragment>
-     
         <FrameLayout
             android:id="@+id/frameLayout"
             android:layout_width="match_parent"
