+++ MountainWalker.Droid/Services/DroidMainActivityService.cs
@@ -2,6 +2,7 @@
 using Android.Gms.Maps;
 using Android.Gms.Maps.Model;
 using MountainWalker.Core.Interfaces;
+using MountainWalker.Droid.Fragments;
 using MountainWalker.Droid.Views;
 using Debug = System.Diagnostics.Debug;
 using DialogFragment = MountainWalker.Droid.Fragments.DialogFragment;
@@ -14,14 +15,14 @@ namespace MountainWalker.Droid.Services
         {
             LatLng coordinate = new LatLng(latitude, longitude);
             CameraUpdate yourLocation = CameraUpdateFactory.NewLatLngZoom(coordinate, 17);
-            MainView.Map.AnimateCamera(yourLocation);
+            HomeFragment.Map.AnimateCamera(yourLocation);
         }
 
         public void SetCurrentLocation(double latitude, double longitude)
         {
             LatLng coordinate = new LatLng(latitude, longitude);
             CameraUpdate yourLocation = CameraUpdateFactory.NewLatLngZoom(coordinate, 17);
-            MainView.Map.AnimateCamera(yourLocation);
+            HomeFragment.Map.AnimateCamera(yourLocation);
         }
 
         public void CloseMainDialog()
